make test work
